update: 更新文件 setvoicemailinfo_fuzzer.cpp

+++ test/fuzztest/setvoicemailinfo_fuzzer/setvoicemailinfo_fuzzer.cpp
@@ -114,7 +114,7 @@ void SetVoiceMailInfo(const uint8_t *data, size_t size)
     int32_t offset = 0;
     std::string mailNumber(reinterpret_cast<const char *>(data), size);
     offset += sizeof(int32_t)
-    offset = (offset > size) ? size : offset;
+    offset = (offset > (int32_t)size) ? (int32_t)size : offset;
     std::string mailName(reinterpret_cast<const char *>(data + offset), size - offset);
     auto mailNameU16 = Str8ToStr16(mailName);
     auto mailNumberU16 = Str8ToStr16(mailNumber);
